fix: linter warnings

+++ src/screen_title.cpp
@@ -19,7 +19,7 @@ void UpdateTitleScreen() {
 void DrawTitleScreen() {
     DrawRectangle(0, 0, GetScreenWidth(), GetScreenHeight(), GREEN);
     Vector2 pos = {20, 10};
-    DrawTextEx(font, "TITLE SCREEN", pos, font.baseSize * 3.0f, 4, DARKGREEN);
+    DrawTextEx(font, "TITLE SCREEN", pos, (float)font.baseSize * 3.0f, 4, DARKGREEN);
     DrawText("PRESS ENTER or TAP to JUMP to GAMEPLAY SCREEN", 120, 220, 20, DARKGREEN);
 }
